# [WIP] Faster and simpler linker for Meteor 3