… tools/isobuild/old-linker.js