@zodern 2023-06-21
When linking the thousands of files added from `import * from '@mui/icons-material'`, it is only 2x times faster (504ms vs 1022ms).
However, this includes the new linker generating a sourcemap for all of the files in node_modules, whereas Meteor never did before because it would have been way too slow.

```
old linker:
modules.js?hash=13fc5c6eed8e703f6ec68b7e5f8d2d1d8e746510:3042 Uncaught TypeError: Cannot read properties of null (reading 'useEffect')
    at Module.useEffect (modules.js?hash=13fc5c6eed8e703f6ec68b7e5f8d2d1d8e746510:3042:21)
    at large-imports.js:9:5

new linker:
react.development.js:1634 Uncaught TypeError: Cannot read properties of null (reading 'useEffect')
    at Module.useEffect (react.development.js:1634:1)
    at large-imports.js:10:5
```