@zodern 2023-06-30
Most of the performance improvement was from replacing the Mozilla source map library with the much faster `@parcel/source-map` library. However, I've found a few situations where it is slower than I want, especially for generating empty source maps.

A few years ago I was experimenting with creating a new source map library that tried a different approach to concatenating source maps. I'm working on finishing it so we can use it in Meteor.

The slowest part of source map libraries is decoding the mappings, storing them, and then encoding them. One interesting thing about the mappings is they are stored as values relative to the previous mapping. My library leverages that fact to avoid half the work source map libraries would normally do.

That allows the new library to be 16x faster at generating empty source maps, and another optimization raises that to 5,300x faster. The new linker generates a large number of empty source maps for files that Meteor previously didn't have source maps for, and this removes almost all downsides to the linker for doing that extra work (there might be later parts of the build process that it would still affect the performance of).

In other cases, where it is concatenating existing source maps, it is usually 4 - 5x faster. It replaces one or two mappings per input source map to ensure the relative positions are correct, and copies over the other mappings as is. It still has to decode all of the mappings to calculate the position of the last mapping. The slowest test I've used is linking 100 copies of jquery-ui, and this reduces the time spent creating the final source map from 2,300ms with @parcel/source-map to less than 500ms.

The only downside I've found so far is the final source map might be larger since it doesn't deduplicate the names array, but it also seems the mappings are sometimes smaller.

> When linking the thousands of files added from import * from '@mui/icons-material', it is only 2x times faster (504ms vs 1022ms).

With the new source map library and optimizations I made to generating the header and footer of each file, this is ~220ms faster.