@zodern 2023-07-24
The new source map library is finished (https://www.npmjs.com/package/@zodern/source-maps). It isn't as fast after fixing all of the bugs in the mappings parser, but it is still at least 3.5x faster when used in the linker.