@dallman2 2024-01-03
Hey @zodern, how close is this to being incorporated into the Meteor 3 beta? I have been pulling the thread on why the build times for my (quite large and cumbersome) Meteor app was so slow to rebuild during development. This PR would be a godsend for the ongoing process of migration to 3.0.